Move V2 implementation out of /default-component subfolder
